move SQLQuery tests to their own module
